Update gender_inference_model_1.py
```markdown
# Issue Resolution Documentation

## Issue Description

When running the script `test_booknlp.py`, I encountered a `UnicodeDecodeError`. This error occurred while the script was processing with the following configuration and outputs:

```plaintext
PS C:\Users\super\Pictures> python .\test_booknlp.py
using device cpu
{'pipeline': 'entity,quote,supersense,event,coref', 'model': 'big'}
--- startup: 26.651 seconds ---
--- spacy: 0.031 seconds ---
--- entities: 0.562 seconds ---
--- quotes: 0.000 seconds ---
--- attribution: 0.000 seconds ---
--- name coref: 0.000 seconds ---
Traceback (most recent call last):
  File "C:\Users\super\Pictures\test_booknlp.py", line 19, in <module>
    booknlp.process(input_file, output_directory, book_id)
  File "C:\Users\super\AppData\Local\Programs\Python\Python310\lib\site-packages\booknlp\booknlp.py", line 17, in process
    self.booknlp.process(inputFile, outputFolder, idd)
  File "C:\Users\super\AppData\Local\Programs\Python\Python310\lib\site-packages\booknlp\english\english_booknlp.py", line 426, in process
    genderEM=GenderEM(tokens=tokens, entities=entities, refs=refs, genders=self.gender_cats, hyperparameterFile=self.gender_hyperparameterFile)
  File "C:\Users\super\AppData\Local\Programs\Python\Python310\lib\site-packages\booknlp\english\gender_inference_model_1.py", line 71, in __init__
    self.read_hyperparams(hyperparameterFile)
  File "C:\Users\super\AppData\Local\Programs\Python\Python310\lib\site-packages\booknlp\english\gender_inference_model_1.py", line 167, in read_hyperparams
    header=file.readline().rstrip()
  File "C:\Users\super\AppData\Local\Programs\Python\Python310\lib\encodings\cp1252.py", line 23, in decode
    return codecs.charmap_decode(input,self.errors,decoding_table)[0]
UnicodeDecodeError: 'charmap' codec can't decode byte 0x9d in position 1701: character maps to <undefined>
```

## Resolution Steps

The issue was resolved by specifying the correct encoding when opening a file within the `gender_inference_model_1.py`. This was necessary due to the presence of characters in the file that were not compatible with the default `cp1252` encoding used by Python on Windows.

### Modified Code

The problematic line in the script:

```python
with open(filename) as file:
```

was replaced with:

```python
with open(filename, 'r', encoding='utf-8') as file:
```

This change ensures that the file is read with UTF-8 encoding, which can handle a broader range of characters, thus preventing the `UnicodeDecodeError`.
```